Cause hare to add a newline to payload messages.
If payloads are delivered to Unix programs through STDIN they will
expect a newline character to terminate the message. While the
originator of the payload might add a newline, it's not impossible
that they will fail to or not understand the need. hare strives to be
a good Unix tool, so the newline is added by default.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>